Take the energy ratio as the Decimal it is
The shared constant is now a Decimal, so the AFCD importer stops laundering
it back through str and uses it directly. The importer keeps its own half-up
quantize: the shipped shard was built that way and re-rounding must not move
a published figure, which is checked across every tenth of a kilojoule to
20000.

Callers still holding floats take the float at the end of the conversion
rather than the start, so the exact division happens first and the lossy step
is where it can be seen.

+++ src/pantry/nutrition_test.py
@@ -12,7 +12,9 @@ from pantry.nutrition import (
 
 def test_kilojoules_are_divided_by_the_published_ratio() -> None:
     """4.184 exactly. The old 0.239006 reciprocal was rounded, and wrong."""
-    assert energy_to_kcal(1000, "kJ") == 1000 / KJ_PER_KCAL
+    # The ratio is a Decimal; a record still holds floats, so this is where
+    # the figure crosses into the lossy domain.
+    assert energy_to_kcal(1000, "kJ") == float(1000 / KJ_PER_KCAL)
     assert energy_to_kcal(4184, "kj") == 1000
     assert energy_to_kcal(239, "kcal") == 239
